ci: fix sign-conversion and usearch warnings under strict GCC/Clang
Two distinct cross-platform regressions surfaced on Linux Clang and
GCC (and would have on macOS Clang once it got past llama.cpp):

1. **Sign-conversion in embedder.cpp:76 and reranker.cpp:49.**
   `std::vector<llama_token> tokens(initial)` with int32_t initial
   triggers -Wsign-conversion (signed → unsigned size_type) on
   Clang/GCC, fatal under our -Werror policy. MSVC let it slide
   silently. Fix: cast to std::size_t at the call site.

2. **usearch internals trip our explicit warning list under GCC.**
   The per-include #pragma block in vector_index.cpp suppressed
   only -Wall / -Wextra / -Wpedantic, but our top-level CMakeLists
   also enables -Wshadow, -Wold-style-cast, -Wdouble-promotion,
   -Wconversion, -Wsign-conversion etc. — none of which are folded
   into -Wall. usearch's templates use plenty of old C-style casts
   and float/double mixing internally, so GCC failed compiling
   vector_index.cpp with dozens of [-Werror=*] errors.

   Clang sidestepped this via `#pragma clang diagnostic ignored
   "-Weverything"`, but GCC has no equivalent kill switch.

   Cleaner fix than expanding the pragma list: mark the
   vectra::usearch INTERFACE library's include directory as SYSTEM.
   The compiler (GCC/Clang/MSVC) then treats every header reached
   through it as out-of-tree and suppresses warnings within them
   uniformly, regardless of which warnings we have enabled in our
   own targets.

Locally re-validated against MSVC (53/53 tests pass). The other
four matrix legs validate the cross-platform fix on the next CI
run.

+++ src/embed/embedder.cpp
@@ -73,7 +73,7 @@ void l2_normalize(std::vector<float>& v) {
                                                 std::string_view text,
                                                 bool add_special) {
     const int32_t initial = static_cast<int32_t>(text.size()) + 32;
-    std::vector<llama_token> tokens(initial);
+    std::vector<llama_token> tokens(static_cast<std::size_t>(initial));
     int32_t n = llama_tokenize(vocab,
                                text.data(),
                                static_cast<int32_t>(text.size()),
